requirements.txt
google-cloud-vision
matplotlib
numpy
scikit-image
